Add expired_project_cleaner
This new hammer checks against portal for if project has an active
allocation, and if not, shelves the instances. This could be expanded to
support more advanced project cleanup. Various arguments allow the
operator to be nice (grace days, ignore projects with pending allocs).

Requires https://github.com/ChameleonCloud/portal/pull/522 for the
portal allocation API. The API token must be set as an environment
variable in portal, which is used by this script.