WIP for scheduler framework
in order to move away from the complexity of N systemd units+timers,
this begins an approach where we configure a single scheduler service,
and in its config file, configure what hammers should run and when.

This will ultimately allow us to run a single, persistent container,
log results to a single place, and expose relevant metrics.

TODO:

1. parse a config file instead of hard-coding arguments
2. from that config file, dynamically load list of hammers to execute
3. configure persistent job store, sqlite is probably easiest
4. add health check endpoint for the scheduler
5. add basic metrics for each configured hammer
6. update alerting to reference these metrics